ModelObserver renamed to LikeableObserver Vertical slicing package architecture
