+++ README.md
@@ -281,13 +281,13 @@ $sortedArticles = Article::orderByDislikesCount('asc')->get();
 
 ### Events
 
-On each like added `\Cog\Laravel\Likeable\Events\ModelWasLiked` event is fired.
+On each like added `\Cog\Laravel\Likeable\Like\Events\ModelWasLiked` event is fired.
 
-On each like removed `\Cog\Laravel\Likeable\Events\ModelWasUnliked` event is fired.
+On each like removed `\Cog\Laravel\Likeable\Like\Events\ModelWasUnliked` event is fired.
 
-On each dislike added `\Cog\Laravel\Likeable\Events\ModelWasDisliked` event is fired.
+On each dislike added `\Cog\Laravel\Likeable\Like\Events\ModelWasDisliked` event is fired.
 
-On each dislike removed `\Cog\Laravel\Likeable\Events\ModelWasUndisliked` event is fired.
+On each dislike removed `\Cog\Laravel\Likeable\Like\Events\ModelWasUndisliked` event is fired.
 
 ### Console commands
 
@@ -361,7 +361,7 @@ To make it you should use container [binding interfaces to implementations](http
 
 ```php
 $this->app->bind(
-    \Cog\Contracts\Likeable\Like::class,
+    \Cog\Contracts\Likeable\Like\Like::class,
     \App\Models\CustomLike::class
 );
 ```
@@ -378,7 +378,7 @@ $this->app->singleton(
 After that your `CustomLike` and `CustomService` classes will be instantiable with helper method `app()`.
 
 ```php
-$model = app(\Cog\Contracts\Likeable\Like::class);
+$model = app(\Cog\Contracts\Likeable\Like\Like::class);
 $service = app(\Cog\Contracts\Likeable\LikeableService::class);
 ```
