--- src/Traits/HasLikes.php
@@ -1,24 +1,0 @@
-<?php
-
-/*
- * This file is part of Laravel Likeable.
- *
- * (c) Anton Komarev <a.komarev@cybercog.su>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Cog\Laravel\Likeable\Traits;
-
-/**
- * Trait HasLikes.
- *
- * @deprecated 3.0
- * @see \Cog\Laravel\Likeable\Traits\Likeable
- * @package Cog\Laravel\Likeable\Traits
- */
-trait HasLikes
-{
-    use Likeable;
-}
