+++ tests/Unit/Events/ModelWasDislikedTest.php
@@ -11,7 +11,7 @@
 
 namespace Cog\Tests\Laravel\Likeable\Unit\Events;
 
-use Cog\Laravel\Likeable\Events\ModelWasDisliked;
+use Cog\Laravel\Likeable\Like\Events\ModelWasDisliked;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
