+++ tests/Unit/Events/ModelWasUndislikedTest.php
@@ -11,7 +11,7 @@
 
 namespace Cog\Tests\Laravel\Likeable\Unit\Events;
 
-use Cog\Laravel\Likeable\Events\ModelWasUndisliked;
+use Cog\Laravel\Likeable\Like\Events\ModelWasUndisliked;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
