+++ tests/Unit/Events/ModelWasUnlikedTest.php
@@ -11,7 +11,7 @@
 
 namespace Cog\Tests\Laravel\Likeable\Unit\Events;
 
-use Cog\Laravel\Likeable\Events\ModelWasUnliked;
+use Cog\Laravel\Likeable\Like\Events\ModelWasUnliked;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
