+++ tests/Unit/Relations/LikeTest.php
@@ -11,7 +11,7 @@
 
 namespace Cog\Tests\Laravel\Likeable\Unit\Relations;
 
-use Cog\Laravel\Likeable\Models\Like;
+use Cog\Laravel\Likeable\Like\Models\Like;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
